auth-proxy/requirements.txt
flask==2.3.3
google-auth==2.23.0
requests==2.31.0
gunicorn==21.2.0
